+++ lib/entitas_ff.dart
@@ -1,7 +1,5 @@
 library entitas_ff;
 
 export 'src/state.dart';
-export 'src/state_extended.dart';
 export 'src/behaviour.dart';
 export 'src/widgets.dart';
-export 'src/widgets_extended.dart';